# Conflicts:
#	src/main/java/project/controller/US340CreateTaskController.java
#	src/main/java/project/ui/console/MainMenuUI.java

+++ src/main/java/project/services/ProjectService.java
@@ -212,7 +212,7 @@ public class ProjectService {
 	 * 
 	 * @return the projectCollaborator created
 	 */
-	public ProjectCollaborator createProjectCollaborator(User user, Project project, int costPerEffort) {
+	public ProjectCollaborator createProjectCollaborator(User user, Project project, double costPerEffort) {
 
 		ProjectCollaborator newProjectCollaborator = new ProjectCollaborator(user, costPerEffort);
 		newProjectCollaborator.setProject(project);
